+++ src/app/components/map-search/map-search.component.html
@@ -1,0 +1,27 @@
+<div id="search-bars-div">
+
+  <div class="search-box">
+
+    <div class="arrow-icon-button">
+      <ion-button class="arrow-button" aria-label="Back" fill="clear" slot="start">
+        <ion-icon name="arrow-back-outline" size="medium" color="medium"></ion-icon></ion-button>
+    </div>
+    <div class="search-bars-container">
+          <ion-searchbar
+            searchIcon="null"
+            class="custom" 
+            placeholder="Choose starting point" 
+            show-clear-button="never" 
+            (ionChange)="onSearchChangeStart($event)">
+          </ion-searchbar>
+    
+        <ion-searchbar
+          searchIcon="null"
+          class="custom" 
+          placeholder="Choose destination point" 
+          show-clear-button="never" 
+          (ionChange)="onSearchChangeDestination($event)">
+        </ion-searchbar>
+    </div>
+  </div>
+</div>
