+++ src/app/components/map-search/map-search.component.scss
@@ -1,0 +1,85 @@
+#container {
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 50%;
+  transform: translateY(-50%);
+  z-index:0;
+}
+
+#container strong {
+  font-size: 20px;
+  line-height: 26px;
+}
+
+#container p {
+  font-size: 16px;
+  line-height: 22px;
+  color: #8c8c8c;
+  margin: 0;
+}
+
+#container a {
+  text-decoration: none;
+}
+
+
+#search-bars-div {
+  display: flex;
+  flex-direction: column; /* Stack elements */
+  justify-content: center;
+  align-items: center;
+  padding: 20px; 
+  position: absolute;
+  top: 90px; 
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 10;
+  width: 100%;
+}
+
+.search-box {
+  display: flex;
+  flex-direction: row;
+  background: white;
+  border-radius: 24px;
+  padding: 5px;
+  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+  width: 70%;
+  max-height: 133px;
+  min-width: 394px; 
+  max-width: 400px; 
+}
+
+
+.arrow-button {
+  --border-radius: 50%;
+  --vertical-align: middle;
+  --padding-start: 5px;
+  --padding-end: 0px;
+}
+
+ion-searchbar.custom {
+  --color: #6d6d6f;
+  --background: transparent;
+  --placeholder-font-weight: 600;
+  ::ng-deep{
+    .searchbar-input-container{            
+        .searchbar-input{ 
+            border-radius: 24px;
+            border: 1px solid #D2D2D2;
+            height: 46px;
+            width: 338px;
+      } 
+    }
+  } 
+}
+
+ion-searchbar {
+  transform: scale(1.0); 
+  margin: 0; 
+  width: 100%;
+  max-height:44px;
+  position: relative; 
+}
+
