+++ src/app/home/home.page.html
@@ -17,33 +17,7 @@
 
   <div id="container">
     <app-google-map></app-google-map>
-    <div id="search-bars-div">
-
-      <div class="search-box">
-
-        <div class="arrow-icon-button">
-          <ion-button class="arrow-button" aria-label="Back" fill="clear" slot="start">
-            <ion-icon name="arrow-back-outline" size="medium" color="medium"></ion-icon></ion-button>
-        </div>
-        <div class="search-bars-container">
-              <ion-searchbar
-                searchIcon="null"
-                class="custom" 
-                placeholder="Choose starting point" 
-                show-clear-button="never" 
-                (ionChange)="onSearchChangeStart($event)">
-              </ion-searchbar>
-        
-            <ion-searchbar
-              searchIcon="null"
-              class="custom" 
-              placeholder="Choose destination point" 
-              show-clear-button="never" 
-              (ionChange)="onSearchChangeDestination($event)">
-            </ion-searchbar>
-        </div>
-      </div>
-    </div>
+    <app-map-search></app-map-search>
 
     <!-- Uncomment these forms if needed for login/signup -->
     <!--
